Dashboard refactor/459 move actions tab (#460)
* moved to new component

* fixed bugs

* formatted
